BackEnd/requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
Django==3.0.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
Jinja2==3.0.3
MarkupSafe==2.0.1
mysqlclient==2.1.0
pdfkit==1.0.0
pycparser==2.20
PyJWT==2.1.0
pyodbc==4.0.32
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
